Fixed some domain problem and updated block controller

+++ src/main/java/edu/miu/cs/cs544/repository/BlockRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BlockRepository extends JpaRepository<Block, Long> {
+public interface BlockRepository extends JpaRepository<Block, Integer> {
     public List<Block> getBlocksByCode(String code);
     public List<Block> getBlocksByName(String name);
     public List<Block> getBlocksBySemester(String semester);
